Split the BulkEditorTable into small components
