--- packages/js/src/bulk-editor/components/index.js
@@ -1,1 +1,0 @@
-export { PlaceholderPage } from "./placeholder-page";
