+++ packages/js/src/bulk-editor/components/table/bulk-editor-table.js
@@ -1,0 +1,88 @@
+import { __ } from "@wordpress/i18n";
+import { Table } from "@yoast/ui-library";
+import { noop } from "lodash";
+import { BulkEditorBody } from "./table-body";
+import { BulkEditorHeader } from "./table-header";
+import { getColumnCount } from "./table-helpers";
+
+/**
+ * The bulk editor selection.
+ *
+ * @typedef {Object} BulkEditorSelection
+ * @property {number[]}  [selectedIds]   IDs of the selected items.
+ * @property {boolean}   [isAllSelected] Whether the header "select all" checkbox is checked.
+ * @property {Function}  [onToggleRow]   Called with an item id when its checkbox is toggled.
+ * @property {Function}  [onToggleAll]   Called when the header "select all" checkbox is toggled.
+ */
+/**
+ * The bulk editor inline-edit seam. Editing is per field and several rows can edit at once: a row's Edit opens
+ * its field-set fields, and each open field is applied (saved on its own) or discarded independently.
+ *
+ * @typedef {Object} BulkEditorEditing
+ * @property {Object}   [editingRows]    Edit state keyed by item id: `{ [id]: { openFields, draft, savingField } }`.
+ * @property {Function} [onStartEdit]    Called with an item id to enter edit mode.
+ * @property {Function} [onChangeField]  Called with `{ id, key, value }` when a field changes.
+ * @property {Function} [onApplyField]   Called with `{ id, key }` to save that field.
+ * @property {Function} [onDiscardField] Called with `{ id, key }` to discard that field's changes.
+ * @property {Function} [onCancelEdit]   Called with an item id to cancel all of its open fields at once.
+ */
+
+/**
+ * The bulk editor data table.
+ *
+ * It renders the items it is given for the active field set (the Search or Social tab). Selection and inline
+ * editing are seams: this component is presentational and drives them through the `selection` and `editing` bags.
+ *
+ * @param {Object}              props             The props.
+ * @param {BulkEditorItem[]}    props.items       The items to render.
+ * @param {FieldSet}            props.fieldSet    The active field set (its `fields` drive the editable columns).
+ * @param {BulkEditorSelection} [props.selection] The selection seam.
+ * @param {BulkEditorEditing}   [props.editing]   The inline-edit seam.
+ * @param {boolean}             [props.isLoading] Whether to render skeleton rows instead of data.
+ *
+ * @returns {JSX.Element} The table.
+ */
+export const BulkEditorTable = ( { items, fieldSet, selection = {}, editing = {}, isLoading = false } ) => {
+	const columnCount = getColumnCount( fieldSet.fields );
+	// Pass the seam handlers as props, so the sub-components can read them without per-field defaults.
+	const selectionState = { selectedIds: [], isAllSelected: false, onToggleRow: noop, onToggleAll: noop, ...selection };
+	const editingState = {
+		editingRows: {},
+		onStartEdit: noop,
+		onChangeField: noop,
+		onApplyField: noop,
+		onDiscardField: noop,
+		onCancelEdit: noop,
+		...editing,
+	};
+
+	return (
+		<>
+			{ /* Announces the loading state to assistive tech (the skeleton itself is aria-hidden). */ }
+			<div role="status" className="yst-sr-only">
+				{ isLoading ? __( "Loading content…", "wordpress-seo" ) : "" }
+			</div>
+			<Table aria-label={ fieldSet.label } aria-busy={ isLoading } className="yst-table-auto sm:yst-table-fixed yst-w-full [&_thead]:!yst-border-t-0 [&_td]:yst-align-top [&_th]:yst-align-top [&_th]:yst-font-medium">
+				<colgroup>
+					<col className="sm:yst-w-[4%]" />
+					<col className="sm:yst-w-[20%]" />
+					{ fieldSet.fields.map( ( field ) => (
+						<col key={ field.key } className={ field.width } />
+					) ) }
+					<col className="sm:yst-w-[5%]" />
+				</colgroup>
+				<BulkEditorHeader fields={ fieldSet.fields } columnCount={ columnCount } selection={ selectionState } isLoading={ isLoading } />
+				<Table.Body>
+					<BulkEditorBody
+						items={ items }
+						fields={ fieldSet.fields }
+						columnCount={ columnCount }
+						selection={ selectionState }
+						editing={ editingState }
+						isLoading={ isLoading }
+					/>
+				</Table.Body>
+			</Table>
+		</>
+	);
+};
