+++ packages/js/src/bulk-editor/components/table/table-helpers.js
@@ -1,0 +1,33 @@
+import { __ } from "@wordpress/i18n";
+
+/**
+ * Maps a post status to a label, or "" for published (no label shown).
+ *
+ * @param {string} status The post status.
+ *
+ * @returns {string} The label, or "" when nothing should be shown.
+ */
+export const getStatusLabel = ( status ) => {
+	switch ( status ) {
+		case "draft":
+			return __( "Draft", "wordpress-seo" );
+		case "pending":
+			return __( "Pending", "wordpress-seo" );
+		case "future":
+			return __( "Scheduled", "wordpress-seo" );
+		default:
+			return "";
+	}
+};
+
+/**
+ * The fixed columns (select + title + actions); the field set supplies the rest, including the focus keyphrase.
+ *
+ * @param {FieldSetField[]} fields The active field set's editable columns.
+ *
+ * @returns {number} The total number of columns, used for full-width rows (loading/empty).
+ */
+export const getColumnCount = ( fields ) => 3 + fields.length;
+
+// The edit state for a row that is not being edited (no open fields, empty draft, nothing saving).
+export const EMPTY_ROW_EDIT = { openFields: [], draft: {}, savingField: null };
